dj_app/requirements.txt
django>=2.0
djangorestframework
psycopg2-binary
dj-database-url